requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
cloudinary==1.38.0
cryptography==42.0.3
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.24
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
oauthlib==3.2.2
pathspec==0.12.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.2.0
